+++ packages/google_maps_places/google_maps_places/example/lib/auto_complete.dart
@@ -175,8 +175,11 @@ class _MyAutoCompleteBodyState extends State<AutoCompleteBody> {
         },
       ),
       CheckboxListTile(
-        title: Text(
-            'Location bias:  (${_locationBias.northeast.latitude}, ${_locationBias.northeast.longitude}), (${_locationBias.southwest.latitude}, ${_locationBias.southwest.longitude})'),
+        title: Text('Location bias: '
+            '(${_locationBias.northeast.latitude}, '
+            '${_locationBias.northeast.longitude}), '
+            '(${_locationBias.southwest.latitude}, '
+            '${_locationBias.southwest.longitude})'),
         checkColor: Colors.white,
         value: _withLocationBias,
         onChanged: (bool? value) {
@@ -187,8 +190,11 @@ class _MyAutoCompleteBodyState extends State<AutoCompleteBody> {
         },
       ),
       CheckboxListTile(
-        title: Text(
-            'Location restriction:  (${_locationRestriction.northeast.latitude}, ${_locationRestriction.northeast.longitude}), (${_locationRestriction.southwest.latitude}, ${_locationRestriction.southwest.longitude})'),
+        title: Text('Location restriction:  '
+            '(${_locationRestriction.northeast.latitude}, '
+            '${_locationRestriction.northeast.longitude}), '
+            '(${_locationRestriction.southwest.latitude}, '
+            '${_locationRestriction.southwest.longitude})'),
         checkColor: Colors.white,
         value: _withLocationRestriction,
         onChanged: (bool? value) {