+++ packages/google_maps_places/google_maps_places_platform_interface/CHANGELOG.md
@@ -1,4 +1,3 @@
-
 # 0.0.1
 
-* TODO: Describe initial release.
+- Initial release of Google Maps Places plugin.